Moving some long running tests to functional
